Add Service Done
Add service is done, just need to write the testing script, also have
done the home page traversing for when you log in. ie, when you log in
as the business owner, his home page opens up, do remeber though that
the current homepage is not the finle look has it has yet to be
approved.